+++ packages/cryptoplease/lib/bl/outgoing_transfers/create_outgoing_transfer_bloc/ft/state.dart
