+++ packages/cryptoplease/lib/presentation/screens/authenticated/send_flow/fungible_token/enter_address/enter_address_screen.dart
@@ -1,4 +1,4 @@
-import 'package:cryptoplease/bl/outgoing_transfers/create_outgoing_transfer_bloc/token/bloc.dart';
+import 'package:cryptoplease/bl/outgoing_transfers/create_outgoing_transfer_bloc/ft/bloc.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/presentation/screens/authenticated/send_flow/common/components/enter_address_input_widget.dart';
 import 'package:cryptoplease/presentation/screens/authenticated/send_flow/common/send_flow_router.dart';
